added bet controller and cathegory table

+++ OnlineAuction.API/Application/Commands/DeleteLotCommand.cs
@@ -14,7 +14,7 @@ namespace OnlineAuction.API.Application.Commands
     {
         public long Id { get; set; }
 
-        public DeleteLotCommand(int id)
+        public DeleteLotCommand(long id)
         {
             Id = id;
         }
